worker/requirements.txt
requests
pillow
python-dotenv
openai
apscheduler
